Moved previous methods to "old" folder
These methods depend on GpuVoxelsVictor. This whole setup was awkward, since voxel grids were all declared in global scope, in order to be available to the gpu_voxels_visualizer. By switching to RViz as a visualizer, the voxelgrids can be kept in proper scope, and can be copied properly. Moving forward, GpuVoxelsVictor will be deprecated